tests: Add tests to ensure parity with EKB address translations
The EKB address translation source cannot be imported directly into
libpdbg as it would require moving the project to C++. So instead the
translation functions have been manually moved into libpdbg. Obviously
this introduces the possibility of differences between the two
implementations.

A test case was added to guard against this possibility by doing a
comparative test with the original implementation. This involves
testing every address offset for every target type and possible index.

As there is no simple way of establishing which addresses and indices
are valid for each target we have to test all of them.

Signed-off-by: Alistair Popple <alistair@popple.id.au>